requirements.txt
# AnomalyVision — Python dependencies
# Install with:  pip install -r requirements.txt

# Deep learning
tensorflow>=2.10.0

# Web application
streamlit>=1.28.0

# Data processing
numpy>=1.23.0
pandas>=1.5.0

# Image / video processing
opencv-python>=4.7.0
Pillow>=9.4.0

# Visualisation
matplotlib>=3.6.0

# Evaluation metrics
scikit-learn>=1.2.0
scipy>=1.10.0
